requirements.txt
astroid==1.6.1
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pylint==1.8.2
python-dotenv==0.7.1
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
